Scheduler using asyncio tasks
